libgeos: Open Source Geometry Engine ('GEOS') C API

Provides the Open Source Geometry Engine ('GEOS') as a C API that can be used to write high-performance C and C++ geometry operations using R as an interface. Headers are provided to make linking to and using these functions from C++ code as easy and as safe as possible. This package contains an internal copy of the 'GEOS' library to guarantee the best possible consistency on multiple platforms.

See also

Author

Maintainer: Dewey Dunnington dewey@fishandwhistle.net (ORCID)

Authors:

GEOS authors

Other contributors:

Martin Davis [contributor, copyright holder]

Benjamin Campbell [contributor, copyright holder]

Tomasz Sowa [contributor, copyright holder]

Christian Kaiser [contributor, copyright holder]

David Skea [contributor, copyright holder]

Daniel Baston [contributor, copyright holder]

Sandro Santilli [contributor, copyright holder]

Mateusz Loskot [contributor, copyright holder]

Paul Ramsey [contributor, copyright holder]

Olivier Devillers [contributor, copyright holder]

Sean Gillies [contributor, copyright holder]

Mika Heiskanen [contributor, copyright holder]

Safe Software Inc. [copyright holder]

Refractions Research Inc. [copyright holder]

Vivid Solutions Inc. [copyright holder]